Refactor this code, and fix a bug while I'm at it.
